Extracting out profile info from Create Account
